+++ targets/TARGET_STM/TARGET_STM32F4/serial_device.c
@@ -1,6 +1,6 @@
 /* mbed Microcontroller Library
  *******************************************************************************
- * Copyright (c) 2014, STMicroelectronics
+ * Copyright (c) 2015, STMicroelectronics
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -27,8 +27,10 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  *******************************************************************************
  */
+
 #include "mbed_assert.h"
 #include "serial_api.h"
+#include "serial_api_hal.h"
 
 #if DEVICE_SERIAL
 
@@ -38,60 +40,19 @@
 #include "PeripheralPins.h"
 #include "mbed_error.h"
 
-#define UART_NUM (5)
-
+#define UART_NUM (8)
 static uint32_t serial_irq_ids[UART_NUM] = {0};
-static UART_HandleTypeDef uart_handlers[UART_NUM];
+UART_HandleTypeDef uart_handlers[UART_NUM];
 
 static uart_irq_handler irq_handler;
 
 int stdio_uart_inited = 0;
 serial_t stdio_uart;
 
-#if DEVICE_SERIAL_ASYNCH
-    #define SERIAL_S(obj) (&((obj)->serial))
-#else
-    #define SERIAL_S(obj) (obj)
-#endif
-
-static void init_uart(serial_t *obj)
-{
-    struct serial_s *obj_s = SERIAL_S(obj);
-    UART_HandleTypeDef *huart = &uart_handlers[obj_s->index];
-    huart->Instance = (USART_TypeDef *)(obj_s->uart);
-
-    huart->Init.BaudRate     = obj_s->baudrate;
-    huart->Init.WordLength   = obj_s->databits;
-    huart->Init.StopBits     = obj_s->stopbits;
-    huart->Init.Parity       = obj_s->parity;
-#if DEVICE_SERIAL_FC
-    huart->Init.HwFlowCtl    = obj_s->hw_flow_ctl;
-#else
-    huart->Init.HwFlowCtl    = UART_HWCONTROL_NONE;
-#endif
-    huart->TxXferCount       = 0;
-    huart->TxXferSize        = 0;
-    huart->RxXferCount       = 0;
-    huart->RxXferSize        = 0;
-
-    if (obj_s->pin_rx == NC) {
-        huart->Init.Mode = UART_MODE_TX;
-    } else if (obj_s->pin_tx == NC) {
-        huart->Init.Mode = UART_MODE_RX;
-    } else {
-        huart->Init.Mode = UART_MODE_TX_RX;
-    }
-
-
-    if (HAL_UART_Init(huart) != HAL_OK) {
-        error("Cannot initialize UART\n");
-    }
-}
-
 void serial_init(serial_t *obj, PinName tx, PinName rx)
 {
     struct serial_s *obj_s = SERIAL_S(obj);
-    
+  
     // Determine the UART to use (UART_1, UART_2, ...)
     UARTName uart_tx = (UARTName)pinmap_peripheral(tx, PinMap_UART_TX);
     UARTName uart_rx = (UARTName)pinmap_peripheral(rx, PinMap_UART_RX);
@@ -100,60 +61,70 @@ void serial_init(serial_t *obj, PinName tx, PinName rx)
     obj_s->uart = (UARTName)pinmap_merge(uart_tx, uart_rx);
     MBED_ASSERT(obj_s->uart != (UARTName)NC);
 
-    // Enable USART clock + switch to SystemClock
-    if (obj_s->uart == UART_1) {
-        __USART1_FORCE_RESET();
-        __USART1_RELEASE_RESET();
-        __USART1_CLK_ENABLE();
-#if defined(RCC_USART1CLKSOURCE_SYSCLK) 
-        __HAL_RCC_USART1_CONFIG(RCC_USART1CLKSOURCE_SYSCLK);
-#endif
-        obj_s->index = 0;
-    }
-#if defined(USART2_BASE)
-    if (obj_s->uart == UART_2) {
-        __USART2_FORCE_RESET();
-        __USART2_RELEASE_RESET();
-        __USART2_CLK_ENABLE();
-#if defined(RCC_USART2CLKSOURCE_SYSCLK)
-        __HAL_RCC_USART2_CONFIG(RCC_USART2CLKSOURCE_SYSCLK);
-#endif
-        obj_s->index = 1;
-    }
-#endif
+    // Enable USART clock
+    switch (obj_s->uart) {
+        case UART_1:
+            __HAL_RCC_USART1_FORCE_RESET();
+            __HAL_RCC_USART1_RELEASE_RESET();
+            __HAL_RCC_USART1_CLK_ENABLE();
+            obj_s->index = 0;
+            break;
+            
+        case UART_2:
+            __HAL_RCC_USART2_FORCE_RESET();
+            __HAL_RCC_USART2_RELEASE_RESET();
+            __HAL_RCC_USART2_CLK_ENABLE();
+            obj_s->index = 1;
+            break;
 #if defined(USART3_BASE)
-    if (obj_s->uart == UART_3) {
-        __USART3_FORCE_RESET();
-        __USART3_RELEASE_RESET();
-        __USART3_CLK_ENABLE();
-#if defined(RCC_USART3CLKSOURCE_SYSCLK)
-        __HAL_RCC_USART3_CONFIG(RCC_USART3CLKSOURCE_SYSCLK);
-#endif
-        obj_s->index = 2;
-    }
+        case UART_3:
+            __HAL_RCC_USART3_FORCE_RESET();
+            __HAL_RCC_USART3_RELEASE_RESET();
+            __HAL_RCC_USART3_CLK_ENABLE();
+            obj_s->index = 2;
+            break;
 #endif
 #if defined(UART4_BASE)
-    if (obj_s->uart == UART_4) {
-        __UART4_FORCE_RESET();
-        __UART4_RELEASE_RESET();
-        __UART4_CLK_ENABLE();
-#if defined(RCC_UART4CLKSOURCE_SYSCLK)
-        __HAL_RCC_UART4_CONFIG(RCC_UART4CLKSOURCE_SYSCLK);
-#endif
-        obj_s->index = 3;
-    }
+        case UART_4:
+            __HAL_RCC_UART4_FORCE_RESET();
+            __HAL_RCC_UART4_RELEASE_RESET();
+            __HAL_RCC_UART4_CLK_ENABLE();
+            obj_s->index = 3;
+            break;
 #endif
 #if defined(UART5_BASE)
-    if (obj_s->uart == UART_5) {
-        __HAL_RCC_UART5_FORCE_RESET();
-        __HAL_RCC_UART5_RELEASE_RESET();
-        __UART5_CLK_ENABLE();
-#if defined(RCC_UART5CLKSOURCE_SYSCLK)
-        __HAL_RCC_UART5_CONFIG(RCC_UART5CLKSOURCE_SYSCLK);
+        case UART_5:
+            __HAL_RCC_UART5_FORCE_RESET();
+            __HAL_RCC_UART5_RELEASE_RESET();
+            __HAL_RCC_UART5_CLK_ENABLE();
+            obj_s->index = 4;
+            break;
 #endif
-        obj_s->index = 4;
+#if defined(USART6_BASE)
+        case UART_6:
+            __HAL_RCC_USART6_FORCE_RESET();
+            __HAL_RCC_USART6_RELEASE_RESET();
+            __HAL_RCC_USART6_CLK_ENABLE();
+            obj_s->index = 5;
+            break;
+#endif
+#if defined(UART7_BASE)
+        case UART_7:
+            __HAL_RCC_UART7_FORCE_RESET();
+            __HAL_RCC_UART7_RELEASE_RESET();
+            __HAL_RCC_UART7_CLK_ENABLE();
+            obj_s->index = 6;
+            break;
+#endif
+#if defined(UART8_BASE)
+        case UART_8:
+            __HAL_RCC_UART8_FORCE_RESET();
+            __HAL_RCC_UART8_RELEASE_RESET();
+            __HAL_RCC_UART8_CLK_ENABLE();
+            obj_s->index = 7;
+            break;
+#endif
     }
-#endif
 
     // Configure the UART pins
     pinmap_pinout(tx, PinMap_UART_TX);
@@ -171,7 +142,7 @@ void serial_init(serial_t *obj, PinName tx, PinName rx)
     obj_s->databits = UART_WORDLENGTH_8B;
     obj_s->stopbits = UART_STOPBITS_1;
     obj_s->parity   = UART_PARITY_NONE;
-
+    
 #if DEVICE_SERIAL_FC
     obj_s->hw_flow_ctl = UART_HWCONTROL_NONE;
 #endif
@@ -191,40 +162,63 @@ void serial_init(serial_t *obj, PinName tx, PinName rx)
 void serial_free(serial_t *obj)
 {
     struct serial_s *obj_s = SERIAL_S(obj);
-      
+    
     // Reset UART and disable clock
-    if (obj_s->uart == UART_1) {
-        __USART1_FORCE_RESET();
-        __USART1_RELEASE_RESET();
-        __USART1_CLK_DISABLE();
-    }
-    if (obj_s->uart == UART_2) {
-        __USART2_FORCE_RESET();
-        __USART2_RELEASE_RESET();
-        __USART2_CLK_DISABLE();
-    }
+    switch (obj_s->index) {
+        case 0:
+            __HAL_RCC_USART1_FORCE_RESET();
+            __HAL_RCC_USART1_RELEASE_RESET();
+            __HAL_RCC_USART1_CLK_DISABLE();
+            break;
+        case 1:
+            __HAL_RCC_USART2_FORCE_RESET();
+            __HAL_RCC_USART2_RELEASE_RESET();
+            __HAL_RCC_USART2_CLK_DISABLE();
+            break;
 #if defined(USART3_BASE)
-    if (obj_s->uart == UART_3) {
-        __USART3_FORCE_RESET();
-        __USART3_RELEASE_RESET();
-        __USART3_CLK_DISABLE();
-    }
+        case 2:
+            __HAL_RCC_USART3_FORCE_RESET();
+            __HAL_RCC_USART3_RELEASE_RESET();
+            __HAL_RCC_USART3_CLK_DISABLE();
+            break;
 #endif
 #if defined(UART4_BASE)
-    if (obj_s->uart == UART_4) {
-        __UART4_FORCE_RESET();
-        __UART4_RELEASE_RESET();
-        __UART4_CLK_DISABLE();
-    }
+        case 3:
+            __HAL_RCC_UART4_FORCE_RESET();
+            __HAL_RCC_UART4_RELEASE_RESET();
+            __HAL_RCC_UART4_CLK_DISABLE();
+            break;
 #endif
 #if defined(UART5_BASE)
-    if (obj_s->uart == UART_5) {
-        __UART5_FORCE_RESET();
-        __UART5_RELEASE_RESET();
-        __UART5_CLK_DISABLE();
-    }
+        case 4:
+            __HAL_RCC_UART5_FORCE_RESET();
+            __HAL_RCC_UART5_RELEASE_RESET();
+            __HAL_RCC_UART5_CLK_DISABLE();
+            break;
 #endif
-
+#if defined(USART6_BASE)
+        case 5:
+            __HAL_RCC_USART6_FORCE_RESET();
+            __HAL_RCC_USART6_RELEASE_RESET();
+            __HAL_RCC_USART6_CLK_DISABLE();
+            break;
+#endif
+#if defined(UART7_BASE)
+        case 6:
+            __HAL_RCC_UART7_FORCE_RESET();
+            __HAL_RCC_UART7_RELEASE_RESET();
+            __HAL_RCC_UART7_CLK_DISABLE();
+            break;
+#endif
+#if defined(UART8_BASE)
+        case 7:
+            __HAL_RCC_UART8_FORCE_RESET();
+            __HAL_RCC_UART8_RELEASE_RESET();
+            __HAL_RCC_UART8_CLK_DISABLE();
+            break;
+#endif
+    }
+    
     // Configure GPIOs
     pin_function(obj_s->pin_tx, STM_PIN_DATA(STM_MODE_INPUT, GPIO_NOPULL, 0));
     pin_function(obj_s->pin_rx, STM_PIN_DATA(STM_MODE_INPUT, GPIO_NOPULL, 0));
@@ -235,41 +229,8 @@ void serial_free(serial_t *obj)
 void serial_baud(serial_t *obj, int baudrate)
 {
     struct serial_s *obj_s = SERIAL_S(obj);
-
+  
     obj_s->baudrate = baudrate;
-    init_uart(obj);
-}
-
-void serial_format(serial_t *obj, int data_bits, SerialParity parity, int stop_bits)
-{
-    struct serial_s *obj_s = SERIAL_S(obj);
-
-    if (data_bits == 9) {
-        obj_s->databits = UART_WORDLENGTH_9B;
-    } else {
-        obj_s->databits = UART_WORDLENGTH_8B;
-    }
-
-    switch (parity) {
-        case ParityOdd:
-            obj_s->parity = UART_PARITY_ODD;
-            break;
-        case ParityEven:
-            obj_s->parity = UART_PARITY_EVEN;
-            break;
-        default: // ParityNone
-        case ParityForced0: // unsupported!
-        case ParityForced1: // unsupported!
-            obj_s->parity = UART_PARITY_NONE;
-            break;
-    }
-
-    if (stop_bits == 2) {
-        obj_s->stopbits = UART_STOPBITS_2;
-    } else {
-        obj_s->stopbits = UART_STOPBITS_1;
-    }
-
     init_uart(obj);
 }
 
@@ -284,20 +245,19 @@ static void uart_irq(int id)
     if (serial_irq_ids[id] != 0) {
         if (__HAL_UART_GET_FLAG(huart, UART_FLAG_TC) != RESET) {
             if (__HAL_UART_GET_IT_SOURCE(huart, UART_IT_TC) != RESET) {
-            irq_handler(serial_irq_ids[id], TxIrq);
-                __HAL_UART_CLEAR_FLAG(huart, UART_CLEAR_TCF);
-        }
+                irq_handler(serial_irq_ids[id], TxIrq);
+                __HAL_UART_CLEAR_FLAG(huart, UART_FLAG_TC);
+            }
         }
         if (__HAL_UART_GET_FLAG(huart, UART_FLAG_RXNE) != RESET) {
             if (__HAL_UART_GET_IT_SOURCE(huart, UART_IT_RXNE) != RESET) {
-            irq_handler(serial_irq_ids[id], RxIrq);
-                volatile uint32_t tmpval = huart->Instance->RDR; // Clear RXNE flag
-                UNUSED(tmpval);
+                irq_handler(serial_irq_ids[id], RxIrq);
+                __HAL_UART_CLEAR_FLAG(huart, UART_FLAG_RXNE);
             }
         }
         if (__HAL_UART_GET_FLAG(huart, UART_FLAG_ORE) != RESET) {
-            if (__HAL_UART_GET_IT_SOURCE(huart, UART_IT_ORE) != RESET) {
-                __HAL_UART_CLEAR_FLAG(huart, UART_CLEAR_OREF);
+            if (__HAL_UART_GET_IT_SOURCE(huart, USART_IT_ERR) != RESET) {
+                volatile uint32_t tmpval = huart->Instance->DR; // Clear ORE flag
             }
         }
     }
@@ -334,6 +294,27 @@ static void uart5_irq(void)
 }
 #endif
 
+#if defined(USART6_BASE)
+static void uart6_irq(void)
+{
+    uart_irq(5);
+}
+#endif
+
+#if defined(UART7_BASE)
+static void uart7_irq(void)
+{
+    uart_irq(6);
+}
+#endif
+
+#if defined(UART8_BASE)
+static void uart8_irq(void)
+{
+    uart_irq(7);
+}
+#endif
+
 void serial_irq_handler(serial_t *obj, uart_irq_handler handler, uint32_t id)
 {
     struct serial_s *obj_s = SERIAL_S(obj);
@@ -349,36 +330,53 @@ void serial_irq_set(serial_t *obj, SerialIrq irq, uint32_t enable)
     IRQn_Type irq_n = (IRQn_Type)0;
     uint32_t vector = 0;
 
-    if (obj_s->uart == UART_1) {
-        irq_n = USART1_IRQn;
-        vector = (uint32_t)&uart1_irq;
-    }
+    switch (obj_s->index) {
+        case 0:
+            irq_n = USART1_IRQn;
+            vector = (uint32_t)&uart1_irq;
+            break;
 
-    if (obj_s->uart == UART_2) {
-        irq_n = USART2_IRQn;
-        vector = (uint32_t)&uart2_irq;
-    }
-
+        case 1:
+            irq_n = USART2_IRQn;
+            vector = (uint32_t)&uart2_irq;
+            break;
 #if defined(USART3_BASE)
-    if (obj_s->uart == UART_3) {
-        irq_n = USART3_IRQn;
-        vector = (uint32_t)&uart3_irq;
-    }
+        case 2:
+            irq_n = USART3_IRQn;
+            vector = (uint32_t)&uart3_irq;
+            break;
 #endif
-
 #if defined(UART4_BASE)
-    if (obj_s->uart == UART_4) {
-        irq_n = UART4_IRQn;
-        vector = (uint32_t)&uart4_irq;
-    }
+        case 3:
+            irq_n = UART4_IRQn;
+            vector = (uint32_t)&uart4_irq;
+            break;
 #endif
-
 #if defined(UART5_BASE)
-    if (obj_s->uart == UART_5) {
-        irq_n = UART5_IRQn;
-        vector = (uint32_t)&uart5_irq;
-    }
+        case 4:
+            irq_n = UART5_IRQn;
+            vector = (uint32_t)&uart5_irq;
+            break;
 #endif
+#if defined(USART6_BASE)
+        case 5:
+            irq_n = USART6_IRQn;
+            vector = (uint32_t)&uart6_irq;
+            break;
+#endif
+#if defined(UART7_BASE)
+        case 6:
+            irq_n = UART7_IRQn;
+            vector = (uint32_t)&uart7_irq;
+            break;
+#endif
+#if defined(UART8_BASE)
+        case 7:
+            irq_n = UART8_IRQn;
+            vector = (uint32_t)&uart8_irq;
+            break;
+#endif
+    }
 
     if (enable) {
         if (irq == RxIrq) {
@@ -386,8 +384,8 @@ void serial_irq_set(serial_t *obj, SerialIrq irq, uint32_t enable)
         } else { // TxIrq
             __HAL_UART_ENABLE_IT(huart, UART_IT_TC);
         }
-        NVIC_SetVector(irq_n, vector);
-        NVIC_EnableIRQ(irq_n);
+            NVIC_SetVector(irq_n, vector);
+            NVIC_EnableIRQ(irq_n);
 
     } else { // disable
         int all_disabled = 0;
@@ -406,7 +404,7 @@ void serial_irq_set(serial_t *obj, SerialIrq irq, uint32_t enable)
         }
 
         if (all_disabled) {
-            NVIC_DisableIRQ(irq_n);
+          NVIC_DisableIRQ(irq_n);
         }
     }
 }
@@ -419,26 +417,18 @@ int serial_getc(serial_t *obj)
 {
     struct serial_s *obj_s = SERIAL_S(obj);
     UART_HandleTypeDef *huart = &uart_handlers[obj_s->index];
-
+    
     while (!serial_readable(obj));
-    if (obj_s->databits == UART_WORDLENGTH_8B) {
-        return (int)(huart->Instance->RDR & (uint8_t)0xFF);
-    } else {
-        return (int)(huart->Instance->RDR & (uint16_t)0x1FF);
-    }
+    return (int)(huart->Instance->DR & 0x1FF);
 }
 
 void serial_putc(serial_t *obj, int c)
 {
     struct serial_s *obj_s = SERIAL_S(obj);
     UART_HandleTypeDef *huart = &uart_handlers[obj_s->index];
-
+    
     while (!serial_writable(obj));
-    if (obj_s->databits == UART_WORDLENGTH_8B) {
-        huart->Instance->TDR = (uint8_t)(c & (uint8_t)0xFF);
-    } else {
-        huart->Instance->TDR = (uint16_t)(c & (uint16_t)0x1FF);
-    }
+    huart->Instance->DR = (uint32_t)(c & 0x1FF);
 }
 
 int serial_readable(serial_t *obj)
@@ -463,9 +453,9 @@ void serial_clear(serial_t *obj)
 {
     struct serial_s *obj_s = SERIAL_S(obj);
     UART_HandleTypeDef *huart = &uart_handlers[obj_s->index];
-
-    __HAL_UART_CLEAR_FLAG(huart, UART_CLEAR_TCF);
-    __HAL_UART_SEND_REQ(huart, UART_RXDATA_FLUSH_REQUEST);
+    
+    huart->TxXferCount = 0;
+    huart->RxXferCount = 0;
 }
 
 void serial_pinout_tx(PinName tx)
@@ -566,14 +556,16 @@ static IRQn_Type serial_get_irq_n(serial_t *obj)
     IRQn_Type irq_n;
 
     switch (obj_s->index) {
+#if defined(USART1_BASE)
         case 0:
             irq_n = USART1_IRQn;
             break;
-
+#endif
+#if defined(USART2_BASE)
         case 1:
             irq_n = USART2_IRQn;
             break;
-
+#endif
 #if defined(USART3_BASE)
         case 2:
             irq_n = USART3_IRQn;
@@ -584,9 +576,24 @@ static IRQn_Type serial_get_irq_n(serial_t *obj)
             irq_n = UART4_IRQn;
             break;
 #endif
-#if defined(UART5_BASE)
+#if defined(USART5_BASE)
         case 4:
             irq_n = UART5_IRQn;
+            break;
+#endif
+#if defined(USART6_BASE)
+        case 5:
+            irq_n = USART6_IRQn;
+            break;
+#endif
+#if defined(UART7_BASE)
+        case 6:
+            irq_n = UART7_IRQn;
+            break;
+#endif
+#if defined(UART8_BASE)
+        case 7:
+            irq_n = UART8_IRQn;
             break;
 #endif
         default:
@@ -595,7 +602,6 @@ static IRQn_Type serial_get_irq_n(serial_t *obj)
     
     return irq_n;
 }
-
 
 /******************************************************************************
  * MBED API FUNCTIONS
@@ -733,22 +739,19 @@ uint8_t serial_rx_active(serial_t *obj)
 
 void HAL_UART_TxCpltCallback(UART_HandleTypeDef *huart) {
     if (__HAL_UART_GET_FLAG(huart, UART_FLAG_TC) != RESET) {
-        __HAL_UART_CLEAR_FLAG(huart, UART_CLEAR_TCF);
+        __HAL_UART_CLEAR_FLAG(huart, UART_FLAG_TC);
     }
 }
 
 void HAL_UART_ErrorCallback(UART_HandleTypeDef *huart) {
     if (__HAL_UART_GET_FLAG(huart, UART_FLAG_PE) != RESET) {
-        __HAL_UART_CLEAR_FLAG(huart, UART_CLEAR_PEF);
-    }
-    if (__HAL_UART_GET_FLAG(huart, UART_FLAG_FE) != RESET) {
-        __HAL_UART_CLEAR_FLAG(huart, UART_CLEAR_FEF);
-    }
-    if (__HAL_UART_GET_FLAG(huart, UART_FLAG_NE) != RESET) {
-        __HAL_UART_CLEAR_FLAG(huart, UART_CLEAR_NEF);
-    }
-    if (__HAL_UART_GET_FLAG(huart, UART_FLAG_ORE) != RESET) {
-        __HAL_UART_CLEAR_FLAG(huart, UART_CLEAR_OREF);
+        volatile uint32_t tmpval = huart->Instance->DR; // Clear PE flag
+    } else if (__HAL_UART_GET_FLAG(huart, UART_FLAG_FE) != RESET) {
+        volatile uint32_t tmpval = huart->Instance->DR; // Clear FE flag
+    } else if (__HAL_UART_GET_FLAG(huart, UART_FLAG_NE) != RESET) {
+        volatile uint32_t tmpval = huart->Instance->DR; // Clear NE flag
+    } else if (__HAL_UART_GET_FLAG(huart, UART_FLAG_ORE) != RESET) {
+        volatile uint32_t tmpval = huart->Instance->DR; // Clear ORE flag
     }
 }
 
@@ -845,8 +848,8 @@ void serial_tx_abort_asynch(serial_t *obj)
     __HAL_UART_DISABLE_IT(huart, UART_IT_TXE);
     
     // clear flags
-    __HAL_UART_CLEAR_FLAG(huart, UART_CLEAR_TCF);
-    
+    __HAL_UART_CLEAR_FLAG(huart, UART_FLAG_TC);
+
     // reset states
     huart->TxXferCount = 0;
     // update handle state
@@ -874,9 +877,8 @@ void serial_rx_abort_asynch(serial_t *obj)
     __HAL_UART_DISABLE_IT(huart, UART_IT_ERR);
     
     // clear flags
-    __HAL_UART_CLEAR_FLAG(huart, UART_CLEAR_PEF | UART_CLEAR_FEF | UART_CLEAR_OREF);
-    volatile uint32_t tmpval = huart->Instance->RDR; // Clear RXNE flag
-    UNUSED(tmpval);
+    __HAL_UART_CLEAR_FLAG(huart, UART_FLAG_RXNE);
+    volatile uint32_t tmpval = huart->Instance->DR; // Clear errors flag
     
     // reset states
     huart->RxXferCount = 0;
